+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## next
-* Added `focusVisible` option to support the `:focus-visible` selector
+* Migrated to `:focus-visible` instead of the old `:focus` selector
 (by Eduard Aksamitov).
+* Added `oldFocus` option to support the old focus (by Eduard Aksamitov).
 
 ## 5.0.1
 * Added funding links.